src/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
twilio==8.10.1
lxml==4.9.3